core-backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg==3.2.10
PyJWT==2.10.1
sqlparse==0.5.3
